+++ packages/core/src/components/viewer-grid/viewer-grid.template.ts
@@ -2,17 +2,21 @@ import { ViewerGrid } from "@dso-toolkit/sources";
 import { html, TemplateResult } from "lit-html";
 
 export function viewerGridTemplate({
+  filterpanel,
   main,
   map,
   overlay,
   noOverlay,
+  filterpanelOpen,
   overlayOpen,
+  filterpanelApply,
+  filterpanelCancel,
   closeOverlay,
 }: ViewerGrid<TemplateResult>) {
   return noOverlay
     ? html`
         <dso-viewer-grid
-          overlay-open=${overlayOpen}
+          ?overlay-open=${overlayOpen}
           @closeOverlay=${closeOverlay}
         >
           <div slot="main">${main}</div>
@@ -21,9 +25,13 @@ export function viewerGridTemplate({
       `
     : html`
         <dso-viewer-grid
-          overlay-open=${overlayOpen}
+          ?filterpanel-open=${filterpanelOpen}
+          ?overlay-open=${overlayOpen}
           @closeOverlay=${closeOverlay}
+          @filterpanelApply=${filterpanelApply}
+          @filterpanelCancel=${filterpanelCancel}
         >
+          <div slot="filterpanel">${filterpanel}</div>
           <div slot="main">${main}</div>
           <div slot="map">${map}</div>
           <div slot="overlay">${overlay}</div>